Add Docker-based multi-GHC testing and long-running subscription example / test
* Add Docker-based multi-GHC testing and long-running subscription example

Testing with GHC 9.10.2 and GHC 9.12.2 and a pure cabal-based build (no Nix).
The purpose is to prepare for the first Hackage release.

Two cabal.project files:

- cabal.project for local builds (refers to packages within current directory structure).
- cabal.project.docker which instructs Cabal to fetch the packages directly from GitHub.

The update-commit.sh script allows one to easily update which commit / branch is used.

The long-running-subscription example exposes a Servant API and is deployable with docker compose.
It comes with a test-api.sh script which verifies that the projection works correctly:

- Queries current state
- Inserts events ; waits two seconds
- Checks that state has been properly updated.

Fixed:
- SQL schema not properly packaged (not specified as extra-source-file).
- Some type signatures added due to simplified sumbsumption for 9.12 compatibility.